requirements.txt
boto3==1.4.4
botocore==1.5.49
docutils==0.13.1
jmespath==0.9.2
olefile==0.44
Pillow==4.1.1
python-dateutil==2.6.0
requests==2.14.2
s3transfer==0.1.10
six==1.10.0
